Checking delete guest

+++ actions/booking_delete_guest.php
@@ -5,6 +5,8 @@ $booking_uid = filter_var($_POST['booking_uid'], FILTER_SANITIZE_NUMBER_INT);
 $guest_uid = filter_var($_POST['guest_uid'], FILTER_SANITIZE_STRING);
 
 $bookingObject = new booking($booking_uid);
-echo $guest_uid;
-$allGuests = $bookingObject->deleteGuest($guest_uid);
+
+if (isset($bookingObject->uid) && isset($guest_uid)) {
+	$allGuests = $bookingObject->deleteGuest($guest_uid);
+}
 ?>
